Refactoring: 디렉토리 수정 (#148)
