+++ backend/src/main/java/com/api/backend/team/controller/TeamParticipantController.java
@@ -2,7 +2,7 @@ package com.api.backend.team.controller;
 
 import static com.api.backend.team.data.ResponseMessage.DELETE_TEAM_PARTICIPANT;
 
-import com.api.backend.global.aop.notify.SendNotify;
+import com.api.backend.notification.aop.annotation.SendNotify;
 import com.api.backend.notification.data.dto.DtoValueExtractor;
 import com.api.backend.team.data.dto.TeamParticipantsDeleteResponse;
 import com.api.backend.team.data.dto.TeamParticipantsDto;
